+++ Salle.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.VisualBasic;
+﻿using iTextSharp.text.pdf.qrcode;
+using Microsoft.VisualBasic;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -18,22 +19,11 @@ namespace AP_CINE_APPLI
         public Salle()
         {
             InitializeComponent();
-            this.FormBorderStyle = FormBorderStyle.None;
-            this.WindowState = FormWindowState.Maximized;
             lblMsg.Text = "";
         }
 
         private void Salle_Load(object sender, EventArgs e)
         {
-            lblMsg.BackColor = Color.White;
-
-            grdSalle.AllowUserToAddRows = false;
-            grdSalle.ReadOnly = true;
-
-            grdSalle.Columns[0].Width = 30;
-            grdSalle.Columns[1].Width = 100;
-
-
             OdbcConnection cnn = new OdbcConnection();
             OdbcCommand cmd = new OdbcCommand();
             OdbcDataReader drr;
@@ -85,7 +75,6 @@ namespace AP_CINE_APPLI
         private bool checkExistSalle(string numsalle)
         {
             lblMsg.Text = "";
-            lblMsg.ForeColor = Color.Black;
             errorProviderNumSalle.SetError(txtNum, "");
             bool existensalle = false;
             int i = 0;
@@ -96,7 +85,6 @@ namespace AP_CINE_APPLI
                     existensalle = true;
                     lblMsg.Text = "Ce numéro de salle existe déjà";
                     errorProviderNumSalle.SetError(txtNum, "Numéro de salle déjà existant");
-                    lblMsg.ForeColor = Color.Red;
                 }
                 else
                 {
@@ -104,6 +92,36 @@ namespace AP_CINE_APPLI
                 }
             }
             return existensalle;
+        }
+
+        private bool salleHasProjection(string mode, string nosalle)
+        {
+            Boolean CanDelete = true;
+
+            OdbcConnection cnn = new OdbcConnection();
+
+            cnn.ConnectionString = varglob.strconnect;
+            cnn.Open();
+
+            OdbcCommand cmd = new OdbcCommand(); OdbcDataReader drr;
+            cmd.CommandText = "select count(nosalle) as nbproj from projection where nosalle ='" + nosalle + "'";
+            cmd.Connection = cnn;
+            drr = cmd.ExecuteReader();
+            drr.Read();
+
+            if (Convert.ToInt16(drr["nbproj"]) > 0)
+            {
+                CanDelete = false;
+                if (mode == "delete" && MessageBox.Show("Attention, la salle que vous allez supprimer possède une projection.\nÊtes-vous sûr de vouloir la supprimer ?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
+                {
+                    CanDelete = true;
+                }
+            }
+
+            drr.Close();
+            cnn.Close();
+
+            return CanDelete;
         }
 
         private void btnAdd_Click(object sender, EventArgs e)
@@ -116,24 +134,23 @@ namespace AP_CINE_APPLI
                 if (!checkExistSalle(txtNum.Text.ToString()))
                 {
                     OdbcConnection cnn = new OdbcConnection();
-                    OdbcCommand cmd = new OdbcCommand();
-
                     cnn.ConnectionString = varglob.strconnect;
                     cnn.Open();
 
+                    OdbcCommand cmd = new OdbcCommand();
                     cmd.CommandText = "insert into salle values ('" + txtNum.Text + "', '" + numCapac.Value + "')";
                     cmd.Connection = cnn;
-                    cmd.ExecuteReader();
+                    cmd.ExecuteNonQuery();
                     cnn.Close();
+
                     lblMsg.Text = "La salle " + txtNum.Text + " a été ajoutée \navec une capacité de " + numCapac.Value + " places";
-                    lblMsg.ForeColor = Color.Blue;
+
                     Salle_Load(sender, e);
                 }
             }
             else
             {
-                lblMsg.Text = "Donnée(s) manquante(s)";
-                lblMsg.ForeColor = Color.Red;
+                lblMsg.Text = "/!\\ Donnée(s) manquante(s) /!\\";
             }
 
         }
@@ -141,46 +158,61 @@ namespace AP_CINE_APPLI
         private void btnEdit_Click(object sender, EventArgs e)
         {
             checkData();
+            lblMsg.Text = "";
 
             if (!string.IsNullOrEmpty(txtNum.Text) && numCapac.Value > 0)
             {
+                string message = "";
+
                 bool unlikeNumSalle = txtNum.Text.ToString() != grdSalle[0, grdSalle.CurrentRow.Index].Value.ToString();
                 bool unlikeCapac = numCapac.Value.ToString() != grdSalle[1, grdSalle.CurrentRow.Index].Value.ToString();
+
+                if (!unlikeNumSalle)
+                {
+                    errorProviderNumSalle.SetError(txtNum, "");
+                }
 
                 OdbcConnection cnn = new OdbcConnection();
 
                 cnn.ConnectionString = varglob.strconnect;
                 cnn.Open();
 
+                if (unlikeNumSalle && !checkExistSalle(txtNum.Text))
+                {
+                    if (salleHasProjection("edit", grdSalle[0, grdSalle.CurrentRow.Index].Value.ToString()))
+                    {
+                        OdbcCommand cmdSalleNum = new OdbcCommand();
+                        cmdSalleNum.CommandText = "update salle set nosalle = '" + txtNum.Text.ToString() + "' where nosalle ='" + grdSalle[0, grdSalle.CurrentRow.Index].Value + "'";
+                        cmdSalleNum.Connection = cnn;
+                        cmdSalleNum.ExecuteNonQuery();
+
+                        message += "\nmodifiée en salle " + txtNum.Text.ToString();
+                    }
+                    else
+                    {
+                        message += "\nLa salle est déjà enregistrer\nsur une projection.\nImpossible de modifier son numéro";
+                    }
+                }
+
                 if (unlikeCapac)
                 {
                     OdbcCommand cmdSalleCapac = new OdbcCommand();
                     cmdSalleCapac.CommandText = "update salle set nbplaces = '" + numCapac.Value.ToString() + "' where nosalle ='" + grdSalle[0, grdSalle.CurrentRow.Index].Value + "'";
                     cmdSalleCapac.Connection = cnn;
-                    cmdSalleCapac.ExecuteReader();
-                }
+                    cmdSalleCapac.ExecuteNonQuery();
 
-
-                if (unlikeNumSalle && !checkExistSalle(txtNum.Text))
-                {
-                    OdbcCommand cmdSalleNum = new OdbcCommand();
-                    cmdSalleNum.CommandText += unlikeNumSalle ? "update salle set nosalle = '" + txtNum.Text.ToString() + "' where nosalle ='" + grdSalle[0, grdSalle.CurrentRow.Index].Value + "';" : "";
-                    cmdSalleNum.Connection = cnn;
-                    cmdSalleNum.ExecuteReader();
+                    message += "\nnouvelle capacité : " + numCapac.Value.ToString();
                 }
 
                 cnn.Close();
 
                 if ((unlikeNumSalle && !checkExistSalle(txtNum.Text)) || unlikeCapac)
                 {
-                    string message = "Salle " + grdSalle[0, grdSalle.CurrentRow.Index].Value.ToString() + " :";
-                    message += unlikeNumSalle ? "\nmodifiée en salle " + txtNum.Text.ToString() : "";
-                    message += unlikeCapac ? "\nnouvelle capacité : " + numCapac.Value : "";
-                    lblMsg.Text = message;
+                    lblMsg.Text = "Salle " + grdSalle[0, grdSalle.CurrentRow.Index].Value.ToString() + " :" + message;
 
                     Salle_Load(sender, e);
                 }
-                
+
             }
             
         }
@@ -188,22 +220,25 @@ namespace AP_CINE_APPLI
         private void btnDelete_Click(object sender, EventArgs e)
         {
             removeError();
-            if (grdSalle.RowCount > 0 && MessageBox.Show("Êtes-vous sûr de vouloir supprimer la salle " + grdSalle[0, grdSalle.CurrentRow.Index].Value + " ?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
+            if (grdSalle.RowCount > 0 && MessageBox.Show("Êtes-vous sûr de vouloir supprimer la salle " + grdSalle[0, grdSalle.CurrentRow.Index].Value + " ?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes && salleHasProjection("delete", grdSalle[0, grdSalle.CurrentRow.Index].Value.ToString()))
             {
                 OdbcConnection cnn = new OdbcConnection();
-
                 cnn.ConnectionString = varglob.strconnect;
                 cnn.Open();
 
-                OdbcCommand cmdsalle = new OdbcCommand();
+                OdbcCommand cmdprojection = new OdbcCommand();
+                cmdprojection.CommandText = "delete from projection where nosalle ='" + grdSalle[0, grdSalle.CurrentRow.Index].Value + "'";
+                cmdprojection.Connection = cnn;
+                cmdprojection.ExecuteNonQuery();
 
+                OdbcCommand cmdsalle = new OdbcCommand();
                 cmdsalle.CommandText = "delete from salle where nosalle ='" + grdSalle[0, grdSalle.CurrentRow.Index].Value + "'";
                 cmdsalle.Connection = cnn;
-                cmdsalle.ExecuteReader();
+                cmdsalle.ExecuteNonQuery();
+
                 cnn.Close();
 
                 lblMsg.Text = "La salle " + grdSalle[0, grdSalle.CurrentRow.Index].Value + " a été supprimée";
-                lblMsg.ForeColor = Color.Red;
 
                 Salle_Load(sender, e);
             }
